客户端登录拦截

+++ src/main/java/com/bbs/controller/manager/TopicManagerController.java
@@ -32,21 +32,21 @@ public class TopicManagerController {
         return topicService.pageQuery(dto);
     }
 
-    @ApiImplicitParam(name = "topicId",value = "帖子id",defaultValue = "1",dataType = "long")
+    @ApiImplicitParam(name = "topicId",value = "帖子id",defaultValue = "1",dataType = "Integer")
     @ApiOperation(value = "帖子置顶")
     @GetMapping(value = "/setTopicToTop")
     public DataResult<String> setTopicToTop(@RequestParam Long topicId){
         return topicService.setTopicToTop(topicId);
     }
 
-    @ApiImplicitParam(name = "topicId",value = "帖子id",defaultValue = "1",dataType = "long")
+    @ApiImplicitParam(name = "topicId",value = "帖子id",defaultValue = "1",dataType = "Integer")
     @ApiOperation(value = "帖子置为热帖")
     @GetMapping(value = "/setTopicToHot")
     public DataResult<String> setTopicToHot(@RequestParam Long topicId){
         return topicService.setTopicToHot(topicId);
     }
 
-    @ApiImplicitParam(name = "topicId",value = "帖子id",defaultValue = "1",dataType = "long")
+    @ApiImplicitParam(name = "topicId",value = "帖子id",defaultValue = "1",dataType = "Integer")
     @ApiOperation(value = "删除帖子")
     @GetMapping(value = "/deleteTopic")
     public DataResult<String> deleteTopic(@RequestParam Long topicId){
